4단계 진입 유저 권한 OK 서비스 권한 수정중

+++ src/main/java/com/example/scheduledevelop/schedule/Controller/ScheduleController.java
@@ -5,6 +5,8 @@ import com.example.scheduledevelop.schedule.dto.ScheduleResponse;
 import com.example.scheduledevelop.schedule.dto.ScheduleSaveRequest;
 import com.example.scheduledevelop.schedule.dto.ScheduleSaveResponse;
 import com.example.scheduledevelop.schedule.dto.ScheduleUpdateRequest;
+import com.example.scheduledevelop.user.dto.SessionUser;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -37,14 +39,27 @@ public class ScheduleController {
     @PutMapping("/{id}")
     public ResponseEntity<ScheduleResponse> update(
             @PathVariable Long id,
-            @RequestBody ScheduleUpdateRequest request) {
+            @RequestBody ScheduleUpdateRequest request,
+            @SessionAttribute(name = "loginUser") SessionUser sessionUser
+    ) {
+        if (sessionUser == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         return ResponseEntity.ok(scheduleService.update(id, request));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(
             @PathVariable Long id,
-            @RequestParam String password) {
+            @RequestParam String password,
+            HttpSession session
+    ) {
+        SessionUser sessionUser = (SessionUser) session.getAttribute("loginUser");
+        if (sessionUser == null) {
+            throw new IllegalStateException("권한이 없습니다.");
+        }
+
         scheduleService.delete(id, password);
         return ResponseEntity.noContent().build();
     }
